+++ packages/core/src/general/constants.ts
@@ -1,4 +1,6 @@
 // indexes for padding arrays
+import { FourSideSizeSpec } from './types'
+
 export const TOP = 0,
     RIGHT = 1,
     BOTTOM = 2,
@@ -11,3 +13,5 @@ export const WIDTH = 0,
 // indexes for size and position arrays
 export const X = 0,
     Y = 1
+
+export const zeroPadding: FourSideSizeSpec = [0, 0, 0, 0]